chore: remove dead files and reorganize project structure
- Delete 10 empty stub files (views/, hooks/)
- Delete unused components: LeftNavbar, Dashboard, TeamConcepts
- Move GeoMap from views/ to components/dashboard/geo/ (proper home)
- Update import path in Map.tsx accordingly
- Remove deno.lock (Deno not used in project)
- Remove empty directories: palette/, report/, views/
